Reduce snarl to a simple chatbot, move plugins elsewhere.
