+++ lib/snarl.js
@@ -122,7 +122,9 @@ Snarl.prototype._interpret = function(message) {
       }
 
       if (x === 'help') {
-        var response = 'There are ' + options.length + ' available triggers: `'+join(options)+'`.  To trigger one, use an `!` anywhere in your message.  For example:\n> Maybe we should get together for a !meetup?  Gosh, that would be fun!';
+        var response = 'There are ' + options.length + ' available triggers: '+join(options.map(function(x) {
+          return '`'+x+'`';
+        }))+'.  To trigger one, use an `!` anywhere in your message.  For example:\n> Maybe we should get together for a !meetup?  Gosh, that would be fun!';
         self.emit('response', response, message);
       }
     });
