+++ plugins/meetups.js
@@ -1,0 +1,3 @@
+module.exports = {
+  meetup: 'Many of us have met in person, usually for drinks.  Join #meetups to coordinate!'
+}
